feat(lint): extract the canonical-envelope page audit into @objectstack/lint; gate cloud-connection's pages
The three-door detector from platform-objects' canonical-expression-envelopes
gate moves to @objectstack/lint beside page-walk.ts as
auditPageExpressionEnvelopes / renderBareExpressionFindings, so every package
shipping raw-literal Page exports can run the same gate. platform-objects'
gate now consumes the shared export (population scan, preconditions, verdict
and shipped-page downgrade control unchanged); cloud-connection gains a thin
gate over its two shipped pages, with @objectstack/lint devDependency and the
anchored vitest source alias. MarketplaceInstalledPage is declared : Page
(type-level only) so export-shape discovery sees it.
